+++ docsig/_utils.py
@@ -28,3 +28,13 @@ def almost_equal(str1: str, str2: str, mini: float, maxi: float) -> bool:
     :return: Boolean result for whether both strings are almost equal.
     """
     return mini < _SequenceMatcher(a=str1, b=str2).ratio() < maxi
+
+
+def vprint(msg: str, verbose: bool = False) -> None:
+    """Print messages only if verbose is true.
+
+    :param msg: Message to print.
+    :param verbose: Whether verbose mode is enabled.
+    """
+    if verbose:
+        print(msg)