+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.45.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- verbose option
 ### Changed
 - short form for version